refactor: common event package
