+++ src/main/java/com/egon/msscbeerservice/beer/listeners/impl/BrewBeerListenerImpl.java
@@ -1,7 +1,7 @@
 package com.egon.msscbeerservice.beer.listeners.impl;
 
-import com.egon.msscbeerservice.beer.dtos.events.BeerEventDto;
-import com.egon.msscbeerservice.beer.dtos.events.NewInventoryEventDto;
+import com.egon.common.events.dtos.LowInventoryBeerEventDto;
+import com.egon.common.events.dtos.NewInventoryEventDto;
 import com.egon.msscbeerservice.beer.listeners.BrewBeerListener;
 import com.egon.msscbeerservice.beer.repositories.BeerRepository;
 import com.egon.msscbeerservice.config.JmsConfig;
@@ -11,13 +11,13 @@ import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.springframework.jms.annotation.JmsListener;
 import org.springframework.jms.core.JmsTemplate;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
 import java.util.Optional;
 
 @Slf4j
 @RequiredArgsConstructor
-@Service
+@Component
 public class BrewBeerListenerImpl implements BrewBeerListener {
 
   private final BeerRepository beerRepository;
@@ -25,7 +25,7 @@ public class BrewBeerListenerImpl implements BrewBeerListener {
 
   @JmsListener(destination = JmsConfig.BREWING_REQUEST_QUEUE)
   @Override
-  public void listen(BeerEventDto eventDto) {
+  public void listen(LowInventoryBeerEventDto eventDto) {
     val beerDto = Optional.of(eventDto.getData()).orElseThrow(IllegalArgumentException::new);
     log.debug("Get beer {} listening queue {}", beerDto.id(), JmsConfig.BREWING_REQUEST_QUEUE);
     val beer = beerRepository.findById(beerDto.id()).orElseThrow(NotFoundException::new);
